WFLY-14452 Refactor key affinity service factory implementation into wildfly-clustering-infinispan-spi module
